Move freighter ESLint config to package
The plan is to manage ESLint configs external to the end user's
repository. The entire configuration will be opaque to them, which is
good because it means changing it in the future is actually possible.

I'm not sure if the way I'm using dependencies will actually work.
I might need to change to a JavaScript file so I can `require` the
packages and pass them directly (like `eslint-plugin-flowtype`).