Add CI command
This replaces the CI bash script, but isn't as fancy... yet. To
facilitate some changes, I've added a feature to the `command` decorator
that checks for a special exit code object type. If it sees one and exit
codes have been enabled, it'll exit with the given code. This is in an
effort to make command handlers more composable.