Compile packages before attempting to run CI
So this is awkward, running `yarn ci` depends on `freighter-scripts`,
which needs to be compiled before it can be symlinked. Because the link
target doesn't exist at the time yarn needs it, it just doesn't symlink.

This problem will only exist for the freighter repo itself.

As a remedy, I compile all the packages and then trigger a re-install
(which is significantly faster the second time) and provides the
`freighter-scripts` command.