Replace decorator exit code with FatalError
I don't need the decorator anymore! Yay for abandoning that bucket of
hacks. Instead of the `exit(...)` utility I'm depending on `FatalError`,
which is a cleaner implementation of the same idea built into `dispute`.

Speaking of which, I'm quickly realizing I shouldn't have forced it to
accept a message string. That should probably be optional.