Remove TypeScript from ESLint settings
I'm out. There's so little advantage to adding any TypeScript
assumptions in this package that I'm just giving up. It isn't worth the
hassle in my other projects. Plus, I'm kinda disheartened on types-in-JS
these days. Especially knowing how fantastic it can be in other
languages, and how far we are from it.

Anyway, now the typescript configs aren't symlinked and ESLint doesn't
make an assumption about the parser.