Move several config files into package.json
This is a habit I've taken up with other projects. Much of the project
configuration is already stored in the package.json, and aggregating it
with other tooling just feels natural. I don't know, maybe I'll hate the
pattern in a month or two. But as the only user of Freighter, I have
many liberties.